MOB-1747: Add default implementations for NTPTooltipDelegate
